FNLib

SLATEC.FNLIB — Module

SLATEC.FNLIB

A collection of routines for evaluating elementary and special functions.

Most of the these routines were written by Wayne Fullerton while at LANL. Some were written by Don Amos of SNLA. There are approximately 63 single precision, 63 double precision and 25 complex user callable elementary and special function routines.

Categories

Intrinsic Functions and Fundamental Functions

Unpack/Pack floating point: r9upak, r9pak

Initialize/Evaluate orthogonal polynomial: inits, csevl

Elementary Functions

Elementary Functions: carg, cbrt, clog10

Relative error exponential: exprel

Relative error logarithm: alnrel, r9ln2r

Trigonometric and Hyperbolic Functions

Tangent, cotangent: ctan; cot

Trig in degrees: sindg, cosdg

Arc Trig: casin, cacos, catan, catan2

Hyperbolic: csinh, ccosh, ctanh

Arc hyperbolic asinh, acosh, atanh; r9atn1

Exponential Integrals and Related Functions

Exponential integral: ei, e1; exint

Logarithmic integral: ali

Gamma Functions and Related Functions

Factorial and Binomial: fac, binom

Gamma: gamma, gamlim, gamr

Log gamma: alngam, algams; r9lgmc; c0lgmc

Incomplete gamma: gami, gamic, gamit

Digamma and Poch: psi, psifn, poch, poch1

Beta: beta, albeta, betai

Error Functions and Fresnel Integrals

Error function: erf, erfc

Dawson's function: daws

Bessel Functions

Bessel functions: besj0, besj1, besy0, besy1

Modified (hyperbolic) Bessel functions: besi0, besi1, besk0, besk1

Modified (hyperbolic) Bessel functions, scaled by an exponential: besi0e, besi1e, besk0e, besk1e

Sequences of Bessel functions: besj, besy, besi, besk; besks, beskes

Bessel Functions of Fractional Order

Airy functions: ai, bi

Exponentially scaled Airy functions: aie, bie

Confluent Hypergeometric Functions

Confluent Hypergeometric: chu

Miscellaneous Functions

Spence dilogarithm: spenc

References

netlib/fn

netlib/vfnlib

netlib/slatec/fundoc.f

TOC

FNLib

TOC

Index

Intrinsic Functions and Fundamental Functions

Elementary Functions

Trigonometric and Hyperbolic Functions

Exponential Integrals and Related Functions

Gamma Functions and Related Functions

Error Functions and Fresnel Integrals

Bessel Functions

Sequences of Bessel functions

Bessel Functions of Fractional Order

Confluent Hypergeometric Functions

Miscellaneous Functions

Index

SLATEC.FNLIB

SLATEC.FNLIB.acosh

SLATEC.FNLIB.ai

SLATEC.FNLIB.aie

SLATEC.FNLIB.albeta

SLATEC.FNLIB.algams

SLATEC.FNLIB.ali

SLATEC.FNLIB.alngam

SLATEC.FNLIB.alnrel

SLATEC.FNLIB.asinh

SLATEC.FNLIB.atanh

SLATEC.FNLIB.besi

SLATEC.FNLIB.besi0

SLATEC.FNLIB.besi0e

SLATEC.FNLIB.besi1

SLATEC.FNLIB.besi1e

SLATEC.FNLIB.besj

SLATEC.FNLIB.besj0

SLATEC.FNLIB.besj1

SLATEC.FNLIB.besk

SLATEC.FNLIB.besk0

SLATEC.FNLIB.besk0e

SLATEC.FNLIB.besk1

SLATEC.FNLIB.besk1e

SLATEC.FNLIB.beskes

SLATEC.FNLIB.besks

SLATEC.FNLIB.besy

SLATEC.FNLIB.besy0

SLATEC.FNLIB.besy1

SLATEC.FNLIB.beta

SLATEC.FNLIB.betai

SLATEC.FNLIB.bi

SLATEC.FNLIB.bie

SLATEC.FNLIB.binom

SLATEC.FNLIB.cacos

SLATEC.FNLIB.carg

SLATEC.FNLIB.casin

SLATEC.FNLIB.catan

SLATEC.FNLIB.catan2

SLATEC.FNLIB.cbrt

SLATEC.FNLIB.ccosh

SLATEC.FNLIB.chu

SLATEC.FNLIB.clog10

SLATEC.FNLIB.cosdg

SLATEC.FNLIB.cot

SLATEC.FNLIB.csevl

SLATEC.FNLIB.csinh

SLATEC.FNLIB.ctan

SLATEC.FNLIB.ctanh

SLATEC.FNLIB.daws

SLATEC.FNLIB.e1

SLATEC.FNLIB.ei

SLATEC.FNLIB.erf

SLATEC.FNLIB.erfc

SLATEC.FNLIB.exint

SLATEC.FNLIB.exprel

SLATEC.FNLIB.fac

SLATEC.FNLIB.gami

SLATEC.FNLIB.gamic

SLATEC.FNLIB.gamit

SLATEC.FNLIB.gamlim

SLATEC.FNLIB.gamma

SLATEC.FNLIB.gamr

SLATEC.FNLIB.inits

SLATEC.FNLIB.poch

SLATEC.FNLIB.poch1

SLATEC.FNLIB.psi

SLATEC.FNLIB.r9atn1

SLATEC.FNLIB.r9lgmc

SLATEC.FNLIB.r9ln2r

SLATEC.FNLIB.r9pak

SLATEC.FNLIB.r9upak

SLATEC.FNLIB.sindg

SLATEC.FNLIB.spenc

Intrinsic Functions and Fundamental Functions

SLATEC.FNLIB.r9upak — Function

r9upak(x::Float32)
r9upak(x::Float64)

Unpack a floating point number X so that X = Y * 2^N, where Y is between 0.5 and 1.0.

x: Input floating point number

Return:

y: Mantissa

n: Exponent

Fortran Name: R9UPAK(X, Y, N), D9UPAK(X, Y, N)

SLATEC.FNLIB.r9pak — Function

r9pak(y::Float32, n::Int32)
r9pak(y::Float64, n::Int32)

Pack a base 2 exponent into a floating point number.

y: Mantissa

n: Exponent

Reutrn

Packed floating point number

Fortran Name: R9PAK(Y, N), D9PAK(Y, N)

SLATEC.FNLIB.inits — Function

inits(os::Vector{Float32}, eta::Float32)
inits(os::Vector{Float64}, eta::Float64)

Determine the number of terms needed in an orthogonal polynomial series so that it meets a specified accuracy.

os: Coefficients in an orthogonal series.

eta: Requested accuracy of the series.

Reutrn

Number of terms needed to meet the accuracy.

Fortran Name: INITS(OS, NOS, ETA), INITDS(OS, NOS, ETA)

SLATEC.FNLIB.csevl — Function

csevl(x::Float32, cs::Vector{Float32})
csevl(x::Float64, cs::Vector{Float64})

Evaluate Chebyshev series.

Formula: summation for i = 1 to n of cs(i)*(2*x)**(i-1)

x: Input value

cs: Coefficients in the Chebyshev series.

n: Number of coefficients in cs.

Reutrn

Value of the Chebyshev series at x.

Fortran Name: CSEVL(X, CS, N), DCSEVL(X, CS, N)

Elementary Functions

SLATEC.FNLIB.carg — Function

carg(z::ComplexF32)

Argument = theta in radians.

Formula: z = |z| * e**(i * theta)

Fortran Name: CARG(Z)

SLATEC.FNLIB.cbrt — Function

cbrt(x::Float32)
cbrt(x::Float64)
cbrt(x::ComplexF32)

Cube root.

Fortran Name: CBRT(X), DCBRT(X), CCBRT(X)

SLATEC.FNLIB.exprel — Function

exprel(x::Float32)
exprel(x::Float64)
exprel(x::ComplexF32)

Relative error exponential from first order.

Formula: ((e**x) -1) / x

Fortran Name: EXPREL(X), DEXPRL(X), CEXPRL(X)

SLATEC.FNLIB.clog10 — Function

clog10(z::ComplexF32)

Common logarithm.

Formula: log to the base 10 of z

Fortran Name: CLOG10(Z)

SLATEC.FNLIB.alnrel — Function

alnrel(x::Float32)
alnrel(x::Float64)
alnrel(x::ComplexF32)

Relative error logarithm.

Formula: ln(1 + x)

Fortran Name: ALNREL(X), DLNREL(X), CLNREL(X)

SLATEC.FNLIB.r9ln2r — Function

r9ln2r(x::Float32)
r9ln2r(x::Float64)
r9ln2r(x::ComplexF32)

Relative error logarithm from second order.

Formula: (ln(1 + x) - x + x**2/2) / x**3

Fortran Name: R9LN2R(X), D9LN2R(X), C9LN2R(X)

Trigonometric and Hyperbolic Functions

SLATEC.FNLIB.ctan — Function

ctan(z::ComplexF32)

Tangent.

Formula: tan z

Fortran Name: CTAN(Z)

SLATEC.FNLIB.cot — Function

cot(x::Float32)
cot(x::Float64)
cot(x::ComplexF32)

Cotangent.

Formula: cot x

Fortran Name: COT(X), DCOT(X), CCOT(X)

SLATEC.FNLIB.sindg — Function

sindg(x::Float32)
sindg(x::Float64)

Sine x in degrees.

Formula: sin((2*pi*x)/360)

Fortran Name: SINDG(X), DSINDG(X)

SLATEC.FNLIB.cosdg — Function

cosdg(x::Float32)
cosdg(x::Float64)

Cosine x in degrees.

Formula: cos((2*pi*x)/360)

Fortran Name: COSDG(X), DCOSDG(X)

SLATEC.FNLIB.casin — Function

casin(z::ComplexF32)

Arc sine.

Formula: arcsin (z)

Fortran Name: CASIN(Z)

SLATEC.FNLIB.cacos — Function

cacos(z::ComplexF32)

Arc cosine.

Formula: arccos (z)

Fortran Name: CACOS(Z)

SLATEC.FNLIB.catan — Function

catan(z::ComplexF32)

Arc tangent.

Formula: arctan (z)

Fortran Name: CATAN(Z)

SLATEC.FNLIB.catan2 — Function

catan2(z1::ComplexF32, z2::ComplexF32)

Quadrant correct arc tangent.

Formula: arctan (z1/z2)

Fortran Name: CATAN2(Z1, Z2)

SLATEC.FNLIB.csinh — Function

csinh(z::ComplexF32)

Hyperbolic sine.

Formula: sinh z

Fortran Name: CSINH(Z)

SLATEC.FNLIB.ccosh — Function

ccosh(z::ComplexF32)

Hyperbolic cosine.

Formula: cosh z

Fortran Name: CCOSH(Z)

SLATEC.FNLIB.ctanh — Function

ctanh(z::ComplexF32)

Hyperbolic tangent.

Formula: tanh z

Fortran Name: CTANH(Z)

SLATEC.FNLIB.asinh — Function

asinh(x::Float32)
asinh(x::Float64)
asinh(x::ComplexF32)

Arc hyperbolic sine.

Formula: arcsinh (x)

Fortran Name: ASINH(X), DASINH(X), CASINH(X)

SLATEC.FNLIB.acosh — Function

acosh(x::Float32)
acosh(x::Float64)
acosh(x::ComplexF32)

Arc hyperbolic cosine.

Formula: arccosh (x)

Fortran Name: ACOSH(X), DACOSH(X), CACOSH(X)

SLATEC.FNLIB.atanh — Function

atanh(x::Float32)
atanh(x::Float64)
atanh(x::ComplexF32)

Arc hyperbolic tangent.

Formula: arctanh (x)

Fortran Name: ATANH(X), DATANH(X), CATANH(X)

SLATEC.FNLIB.r9atn1 — Function

r9atn1(x::Float32)
r9atn1(x::Float64)

Relative error arc tangent from first order.

Formula: (arctan (x) - x) / x**3

Fortran Name: R9ATN1(X), D9ATN1(X)

Exponential Integrals and Related Functions

SLATEC.FNLIB.ei — Function

ei(x::Float32)
ei(x::Float64)

Exponential integral Ei(x).

Formula: (minus) the integral from -x to infinity of (e**-t / t)dt

Fortran Name: EI(X), DEI(X)

SLATEC.FNLIB.e1 — Function

e1(x::Float32)
e1(x::Float64)

Exponential integral E1(x).

Formula: the integral from x to infinity of (e**-t / t) dt

Fortran Name: E1(X), DE1(X)

SLATEC.FNLIB.ali — Function

ali(x::Float32)
ali(x::Float64)

Logarithmic integral li(x).

Formula: the integral from 0 to x of (1 / ln t) dt

Fortran Name: ALI(X), DLI(X)

SLATEC.FNLIB.exint — Function

exint(x::Float32, n::Int32, kode::Int32, m::Int32, tol::Float32)
exint(x::Float64, n::Int32, kode::Int32, m::Int32, tol::Float64)

Compute an m member sequence of Exponential integral: $E_{n+k}(x), k = 0,1,...,(m-1)$ for n >= 1 and x >= 0.

Formula: the integral from 1 to infinity of $(e^{-x*t}/t^{n+k}) dt$

x: x > 0 for n=1 and x >= 0 for n >=2

n: Order, n >= 1

kode: Selection parameter (1=normal, 2=scaled)

m: Number of terms to compute, m >= 1

tol: Relative accuracy wanted, ETOL=R1MACH(4) < tol < 0.1

Return (en, nz, ierr)

en: Result array

nz: Underflow indicator

ierr: Error flag

Fortran Name: EXINT(X,N,KODE,M,TOL, EN,NZ,IERR), DEXINT(X,N,KODE,M,TOL, EN,NZ,IERR)

Gamma Functions and Related Functions

SLATEC.FNLIB.fac — Function

fac(n::Int32) :: Float32
fac(n::Int32, ::Type{Float32}) :: Float32
fac(n::Int32, ::Type{Float64}) :: Float64

Factorial.

Formula: n!

Fortran Name: FAC(N), DFAC(N)

SLATEC.FNLIB.binom — Function

binom(n::Int32, m::Int32) :: Float32
binom(n::Int32, m::Int32, ::Type{Float32}) :: Float32
binom(n::Int32, m::Int32, ::Type{Float64}) :: Float64

Binomial.

Formula: n!/(m!*(n-m)!)

Fortran Name: BINOM(N, M), DBINOM(N, M)

SLATEC.FNLIB.gamma — Function

gamma(x::Float32)
gamma(x::Float64)
gamma(x::ComplexF32)

Gamma.

Formula: gamma(x)

Fortran Name: GAMMA(X), DGAMMA(X), CGAMMA(X)

SLATEC.FNLIB.gamlim — Function

gamlim(::Type{Float32}) :: Tuple{Float32, Float32}
gamlim(::Type{Float64}) :: Tuple{Float64, Float64}

Gamma(x) under and overflow limits.

Return:

xmin: Minimum legal value of X in gamma(X)

xmax: Maximum legal value of X in gamma(X)

Fortran Name: GAMLIM(XMIN, XMAX), DGAMLM(XMIN, XMAX)

SLATEC.FNLIB.gamr — Function

gamr(x::Float32)
gamr(x::Float64)
gamr(x::ComplexF32)

Reciprocal gamma.

Formula: 1 / gamma(x)

Fortran Name: GAMR(X), DGAMR(X), CGAMR(X)

SLATEC.FNLIB.alngam — Function

alngam(x::Float32)
alngam(x::Float64)
alngam(z::ComplexF32)

Log abs gamma.

Formula: ln |gamma(x)|

Fortran Name: ALNGAM(X), DLNGAM(X)

SLATEC.FNLIB.algams — Function

algams(x::Float32)
algams(x::Float64)

Log abs gamma with sign.

Formula: g = ln |gamma(x)|, s = sign gamma(x)

x: Argument

Return:

g: Log absolute gamma

s: Sign of gamma, +1.0 or -1.0

Fortran Name: ALGAMS(X, G, S), DLGAMS(X, G, S)

SLATEC.FNLIB.gami — Function

gami(a::Float32, x::Float32)
gami(a::Float64, x::Float64)

Incomplete gamma.

Formula: gamma(a,x) = integral from 0 to x of (t**(a-1) * e**-t)dt

Fortran Name: GAMI(A, X), DGAMI(A, X)

SLATEC.FNLIB.gamic — Function

gamic(a::Float32, x::Float32)
gamic(a::Float64, x::Float64)

Complementary incomplete gamma.

Formula: gamma(a,x) = integral from x to infinity of (t**(a-1) * e**-t)dt

Fortran Name: GAMIC(A, X), DGAMIC(A, X)

SLATEC.FNLIB.gamit — Function

gamit(a::Float32, x::Float32)
gamit(a::Float64, x::Float64)

Tricomi's incomplete gamma.

Formula: gamma*(a,x) = x**-a * incomplete gamma(a,x) / gamma(a)

Fortran Name: GAMIT(A, X), DGAMIT(A, X)

SLATEC.FNLIB.psi — Function

psi(x::Float32)
psi(x::Float64)
psi(x::ComplexF32)

Psi (Digamma).

Formula: psi(x) = gamma'(x) / gamma(x)

Fortran Name: PSI(X), DPSI(X), CPSI(X)

SLATEC.FNLIB.poch — Function

poch(a::Float32, x::Float32)
poch(a::Float64, x::Float64)

Pochhammer's generalized symbol.

Formula: (a) sub x = gamma(a+x) / gamma(a)

Fortran Name: POCH(A, X), DPOCH(A, X)

SLATEC.FNLIB.poch1 — Function

poch1(a::Float32, x::Float32)
poch1(a::Float64, x::Float64)

Pochhammer's symbol from first order.

Formula: ((a) sub x -1) / x

Fortran Name: POCH1(A, X), DPOCH1(A, X)

SLATEC.FNLIB.beta — Function

beta(a::Float32, b::Float32)
beta(a::Float64, b::Float64)
beta(a::ComplexF32, b::ComplexF32)

Beta.

Formula: b(a,b) = (gamma(a) * gamma(b)) / gamma(a+b) = integral from 0 to 1 of (t(a-1) * (1-t)(b-1))dt

Fortran Name: BETA(A, B), DBETA(A, B), CBETA(A, B)

SLATEC.FNLIB.albeta — Function

albeta(a::Float32, b::Float32)
albeta(a::Float64, b::Float64)
albeta(a::ComplexF32, b::ComplexF32)

Log beta.

Formula: ln b(a,b)

Fortran Name: ALBETA(A, B), DLBETA(A, B), CLBETA(A, B)

SLATEC.FNLIB.betai — Function

betai(x::Float32, a::Float32, b::Float32)
betai(x::Float64, a::Float64, b::Float64)

Incomplete beta.

Formula: i sub x (a,b) = b sub x (a,b) / b(a,b) = 1 / b(a,b) * integral from 0 to x of (t(a-1) * (1-t)(b-1))dt Fortran Name: BETAI(X, A, B), DBETAI(X, A, B)

SLATEC.FNLIB.r9lgmc — Function

r9lgmc(x::Float32)
r9lgmc(x::Float64)
r9lgmc(x::ComplexF32)

Log gamma correction term.

Formula: ln gamma(x) - (ln(2 * pi))/2 - (x - 1/2) * ln(x) + x

Fortran Name: R9LGMC(X), D9LGMC(X), C9LGMC(X)

Error Functions and Fresnel Integrals

SLATEC.FNLIB.erf — Function

erf(x::Float32)
erf(x::Float64)

Error function.

Formula: erf x = (2 / square root of pi) * the integral from 0 to x of e**(-t**2)dt

Fortran Name: ERF(X), DERF(X)

SLATEC.FNLIB.erfc — Function

erfc(x::Float32)
erfc(x::Float64)

Complementary error function.

Formula: erfc x = (2 / square root of pi) * the integral from x to infinity of e**(-t**2)dt

Fortran Name: ERFC(X), DERFC(X)

SLATEC.FNLIB.daws — Function

daws(x::Float32)
daws(x::Float64)

Dawson's function.

Formula: F(x) = e**(-x**2) * the integral from 0 to x of e**(t**2)dt

Fortran Name: DAWS(X), DDAWS(X)

Bessel Functions

SLATEC.FNLIB.besj0 — Function

besj0(x::Float32)
besj0(x::Float64)

Bessel function of the first kind, order zero.

Formula: J sub 0 (x)

Fortran Name: BESJ0(X), DBESJ0(X)

SLATEC.FNLIB.besj1 — Function

besj1(x::Float32)
besj1(x::Float64)

Bessel function of the first kind, order one.

Formula: J sub 1 (x)

Fortran Name: BESJ1(X), DBESJ1(X)

SLATEC.FNLIB.besy0 — Function

besy0(x::Float32)
besy0(x::Float64)

Bessel function of the second kind, order zero.

Formula: Y sub 0 (x)

Fortran Name: BESY0(X), DBESY0(X)

SLATEC.FNLIB.besy1 — Function

besy1(x::Float32)
besy1(x::Float64)

Bessel function of the second kind, order one.

Formula: Y sub 1 (x)

Fortran Name: BESY1(X), DBESY1(X)

SLATEC.FNLIB.besi0 — Function

besi0(x::Float32)
besi0(x::Float64)

Modified Bessel function of the first kind, order zero.

Formula: I sub 0 (x)

Fortran Name: BESI0(X), DBESI0(X)

SLATEC.FNLIB.besi1 — Function

besi1(x::Float32)
besi1(x::Float64)

Modified Bessel function of the first kind, order one.

Formula: I sub 1 (x)

Fortran Name: BESI1(X), DBESI1(X)

SLATEC.FNLIB.besk0 — Function

besk0(x::Float32)
besk0(x::Float64)

Modified Bessel function of the third kind, order zero.

Formula: K sub 0 (x)

Fortran Name: BESK0(X), DBESK0(X)

SLATEC.FNLIB.besk1 — Function

besk1(x::Float32)
besk1(x::Float64)

Modified Bessel function of the third kind, order one.

Formula: K sub 1 (x)

Fortran Name: BESK1(X), DBESK1(X)

SLATEC.FNLIB.besi0e — Function

besi0e(x::Float32)
besi0e(x::Float64)

Modified Bessel function of the first kind, order zero, scaled.

Formula: e**-|x| * I sub 0(x)

Fortran Name: BESI0E(X), DBSI0E(X)

SLATEC.FNLIB.besi1e — Function

besi1e(x::Float32)
besi1e(x::Float64)

Modified Bessel function of the first kind, order one, scaled.

Formula: e**-|x| * I sub 1(x)

Fortran Name: BESI1E(X), DBSI1E(X)

SLATEC.FNLIB.besk0e — Function

besk0e(x::Float32)
besk0e(x::Float64)

Modified Bessel function of the third kind, order zero, scaled.

Formula: e**x * K sub 0 (x)

Fortran Name: BESK0E(X), DBSK0E(X)

SLATEC.FNLIB.besk1e — Function

besk1e(x::Float32)
besk1e(x::Float64)

Modified Bessel function of the third kind, order one, scaled.

Formula: e**x * K sub 1 (x)

Fortran Name: BESK1E(X), DBSK1E(X)

Sequences of Bessel functions

SLATEC.FNLIB.besj — Function

besj(x::Float32, alpha::Float32, n::Int32)
besj(x::Float64, alpha::Float64, n::Int32)

Sequence of Bessel functions of the first kind.

Formula: $J_{a+k-1}(x), k = 1,2,...,N$

x: x >= 0

alpha: Order of first member of the sequence, alpha >= 0

n: Number of members in the sequence, n >= 1

Return (y, nz)

y: Result array

nz: Number of underflows

Fortran Name: BESJ(X,ALPHA,N, Y(*),NZ), DBESJ(X,ALPHA,N, Y(*),NZ)

SLATEC.FNLIB.besy — Function

besy(x::Float32, fnu::Float32, n::Int32)
besy(x::Float64, fnu::Float64, n::Int32)

Sequence of Bessel functions of the second kind.

Formula: $Y_{a+k-1}(x), k = 1,2,...,N$

x: x >= 0

fnu: Order of first member of the sequence, fnu >= 0

n: Number of members in the sequence, n >= 1

Return

y: Result array

Fortran Name: BESY(X,FNU,N, Y(*)), DBESY(X,FNU,N, Y(*))

SLATEC.FNLIB.besi — Function

besi(x::Float32, alpha::Float32, kode::Int32, n::Int32)
besi(x::Float64, alpha::Float64, kode::Int32, n::Int32)

Sequence of modified Bessel functions of the first kind.

Formula: $I_{a+k-1}(x), k = 1,2,...,N$. Optional scaling by exp(-x).

x: x >= 0

alpha: Order of first member of the sequence, alpha >= 0

kode: Selection parameter (1=unscaled, 2=scaled)

n: Number of members in the sequence, n >= 1

Return (y, nz)

y: Result array

nz: Number of underflows

Fortran Name: BESI(X,ALPHA,KODE,N, Y(*),NZ), DBESI(X,ALPHA,KODE,N, Y(*),NZ)

SLATEC.FNLIB.besk — Function

besk(x::Float32, fnu::Float32, kode::Int32, n::Int32)
besk(x::Float64, fnu::Float64, kode::Int32, n::Int32)

Sequence of modified Bessel functions of the third kind.

Formula: $K_{a+k-1}(x), k = 1,2,...,N$. Optional scaling by exp(x).

x: x >= 0

fnu: Order of first member of the sequence, fnu >= 0

kode: Selection parameter (1=unscaled, 2=scaled)

n: Number of members in the sequence, n >= 1

Return (y, nz)

y: Result array

nz: Number of underflows

Fortran Name: BESK(X,FNU,KODE,N, Y(*),NZ), DBESK(X,FNU,KODE,N, Y(*),NZ)

SLATEC.FNLIB.besks — Function

besks(xnu::Float32, x::Float32, n::Int32)
besks(xnu::Float64, x::Float64, n::Int32)

Sequence of modified Bessel functions of the third kind.

Formula: $K_{a+i}(x), i = 0, 1, ..., N-1$ for N > 0 or i = 0, -1, ..., N+1 for N < 0.

xnu: Order of first member of the sequence, xnu >= 0

x: Argument, x >= 0

n: Number of members in the sequence, n >= 1

Return

bk: Result array

Fortran Name: BESKS(XNU,X,N, BK(*)), DBESKS(XNU,X,N, BK(*))

SLATEC.FNLIB.beskes — Function

beskes(xnu::Float32, x::Float32, n::Int32)
beskes(xnu::Float64, x::Float64, n::Int32)

Sequence of modified Bessel functions of the third kind, scaled.

Formula: $e^x * K_{a+i}(x), i = 0, 1, ..., N-1$ for N > 0 or i = 0, -1, ..., N+1 for N < 0.

xnu: Order of first member of the sequence, xnu >= 0

x: Argument, x >= 0

n: Number of members in the sequence, n >= 1

Return

bk: Result array

Fortran Name: BESKES(XNU,X,N, BK(*)), DBSKES(XNU,X,N, BK(*))

Bessel Functions of Fractional Order

SLATEC.FNLIB.ai — Function

ai(x::Float32)
ai(x::Float64)

Airy function Ai.

Formula: Ai(x)

Fortran Name: AI(X), DAI(X)

SLATEC.FNLIB.bi — Function

bi(x::Float32)
bi(x::Float64)

Airy function Bi.

Formula: Bi(x)

Fortran Name: BI(X), DBI(X)

SLATEC.FNLIB.aie — Function

aie(x::Float32)
aie(x::Float64)

Exponentially scaled Airy function Ai.

Formula: Ai(x), x <= 0; exp(2/3 * x**(3/2)) * Ai(x), x >= 0

Fortran Name: AIE(X), DAIE(X)

SLATEC.FNLIB.bie — Function

bie(x::Float32)
bie(x::Float64)

Exponentially scaled Airy function Bi.

Formula: Bi(x), x <= 0; exp(-2/3 * x**(3/2)) * Bi(x), x >= 0

Fortran Name: BIE(X), DBIE(X)

Confluent Hypergeometric Functions

SLATEC.FNLIB.chu — Function

chu(a::Float32, b::Float32, x::Float32)
chu(a::Float64, b::Float64, x::Float64)

Compute the logarithmic confluent hypergeometric function, U(a,b,x)

Crash warning

If input argument do not meet the requirements, the routine will crash julia process!!!

x must be positive, x > 0

This routine is not valid when 1+A-B is close to zero if X is small.

Fortran Name: CHU(A, B, X), DCHU(A, B, X)

Miscellaneous Functions

SLATEC.FNLIB.spenc — Function

spenc(x::Float32)
spenc(x::Float64)

Compute a form of Spence's integral due to K. Mitchell. Also known as the dilogarithm, $Li_2(x)$.

Formula: s(x) = - the integral from 0 to x of ((ln |1-y|) / y)dy

Fortran Name: SPENC(X), DSPENC(X)